requirements.txt
google-generativeai
google-auth
requests  # Optional, only if needed